# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: RailMind Backend
  version: 1.0.0
paths:
...
  /api/v1/stations:
    get:
      tags:
      - operational
      summary: Get Stations
      operationId: get_stations_api_v1_stations_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Page Size
      - name: is_junction
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Junction
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_StationView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/stations/{station_id}:
    get:
      tags:
      - operational
      summary: Get Station
      operationId: get_station_api_v1_stations__station_id__get
      parameters:
      - name: station_id
        in: path
        required: true
        schema:
          type: string
          title: Station Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_StationView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sections:
    get:
      tags:
      - operational
      summary: Get Sections
      operationId: get_sections_api_v1_sections_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Page Size
      - name: corridor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Corridor Id
      - name: section_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Section Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_SectionView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sections/{section_id}:
    get:
      tags:
      - operational
      summary: Get Section
      operationId: get_section_api_v1_sections__section_id__get
      parameters:
      - name: section_id
        in: path
        required: true
        schema:
          type: string
          title: Section Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_SectionView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/train-movements:
    get:
      tags:
      - operational
      summary: Get Train Movements
      operationId: get_train_movements_api_v1_train_movements_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Page Size
      - name: train_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Train Id
      - name: section_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Section Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_TrainMovementView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/disruptions:
    get:
      tags:
      - operational
      summary: Get Disruptions
      operationId: get_disruptions_api_v1_disruptions_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Page Size
      - name: section_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Section Id
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_DisruptionView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/disruptions/{scenario_id}:
    get:
      tags:
      - operational
      summary: Get Disruption
      operationId: get_disruption_api_v1_disruptions__scenario_id__get
      parameters:
      - name: scenario_id
        in: path
        required: true
        schema:
          type: string
          title: Scenario Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DisruptionView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiError:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        httpStatus:
          type: integer
          title: Httpstatus
      type: object
      required:
      - code
      - message
      - httpStatus
      title: ApiError
...
    ApiListResponse_DisruptionView_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/DisruptionView'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[DisruptionView]
...
    ApiListResponse_SectionView_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/SectionView'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[SectionView]
    ApiListResponse_StationView_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/StationView'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[StationView]
...
    ApiListResponse_TrainMovementView_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/TrainMovementView'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[TrainMovementView]
...
    ApiMeta:
      properties:
        timestamp:
          type: string
          title: Timestamp
        requestId:
          type: string
          title: Requestid
        version:
          type: string
          title: Version
      type: object
      required:
      - timestamp
      - requestId
      - version
      title: ApiMeta
...
    ApiResponse_DisruptionView_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/DisruptionView'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[DisruptionView]
...
    ApiResponse_SectionView_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/SectionView'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[SectionView]
...
    ApiResponse_StationView_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/StationView'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[StationView]
...
    DisruptionView:
      properties:
        scenario_id:
          type: string
          title: Scenario Id
        type:
          type: string
          title: Type
        section_id:
          type: string
          title: Section Id
        related_task_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Related Task Id
        planned_start:
          anyOf:
          - type: string
          - type: 'null'
          title: Planned Start
        planned_end:
          anyOf:
          - type: string
          - type: 'null'
          title: Planned End
        actual_end:
          anyOf:
          - type: string
          - type: 'null'
          title: Actual End
        severity:
          type: string
          title: Severity
        description:
          type: string
          title: Description
      type: object
      required:
      - scenario_id
      - type
      - section_id
      - severity
      - description
      title: DisruptionView
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginationMeta:
      properties:
        totalItems:
          type: integer
          title: Totalitems
        page:
          type: integer
          title: Page
        pageSize:
          type: integer
          title: Pagesize
        totalPages:
          type: integer
          title: Totalpages
      type: object
      required:
      - totalItems
      - page
      - pageSize
      - totalPages
      title: PaginationMeta
...
    SectionView:
      properties:
        section_id:
          type: string
          title: Section Id
        from_station_id:
          type: string
          title: From Station Id
        to_station_id:
          type: string
          title: To Station Id
        length_km:
          type: number
          title: Length Km
        track_count:
          type: integer
          title: Track Count
        electrified:
          type: boolean
          title: Electrified
        max_speed_kmph:
          type: integer
          title: Max Speed Kmph
        headway_min:
          type: integer
          title: Headway Min
        section_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Section Type
          default: Main_Line
      type: object
      required:
      - section_id
      - from_station_id
      - to_station_id
      - length_km
      - track_count
      - electrified
      - max_speed_kmph
      - headway_min
      title: SectionView
...
    StationView:
      properties:
        station_id:
          type: string
          title: Station Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        sequence:
          type: integer
          title: Sequence
        km:
          type: number
          title: Km
        category:
          type: string
          title: Category
        platforms:
          type: integer
          title: Platforms
        loop_lines:
          type: integer
          title: Loop Lines
        is_junction:
          type: boolean
          title: Is Junction
      type: object
      required:
      - station_id
      - code
      - name
      - sequence
      - km
      - category
      - platforms
      - loop_lines
      - is_junction
      title: StationView
...
    TrainMovementView:
      properties:
        movement_id:
          type: string
          title: Movement Id
        train_id:
          type: string
          title: Train Id
        section_id:
          type: string
          title: Section Id
        entry_time:
          type: string
          title: Entry Time
        exit_time:
          type: string
          title: Exit Time
        status:
          type: string
          title: Status
        delay_minutes:
          type: integer
          title: Delay Minutes
      type: object
      required:
      - movement_id
      - train_id
      - section_id
      - entry_time
      - exit_time
      - status
      - delay_minutes
      title: TrainMovementView
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...